refactor newtab settings into individual components (and move to a new file)
